Allow --help to work without config, better error
So me, being the gung-ho type that I am, downloaded this and ran it immediately without bothering to read the docs. This resulted in the old fashioned "[] on a nil class" error. I thought something was broken, so I tried running the program with --help, which gave the same problem. This is since the "config" var is only defined if a file exists (which causes Trollop to break when trying to parse the default options).

This changes the example into from a block comment into inline YAML, and changes the URL, causing an HTTP error that will clue the user in that something's amiss if they just run the app, indicating that they need to configure it (and as a bonus, allows --help to work even if the config file doesn't exist).